Icons for CF entities
Such as space, app, org. Includes code that makes it easier to inline
icons rather then the block layout as done on cg-site. To inline an
icon, which is done on dashboard overview page, simply don't use the
icon container and place the icon within the text tag it's next to.

The inlined icon will keep the size of the text tag it's in as it's
height and width are set to 1em. The icon also has a min width/height
though as I found at too small of sizes of text the icon wasn't
rendering very well.

Some of these icons need additional work to receive the border they
have.

This PR addresses separating out fill icons vs regular (or existing)
stroke icons. This has to be done because some of our icons only work
as fills and vice-versa for strokes. The default is just the `icon`
class and is stroke. The `icon-fill` class is for fill icons and will
be colored differently.